backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-environ==0.10.0
django-rosetta==0.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
flake8==6.0.0
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pep8==1.7.1
platformdirs==3.5.1
polib==1.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
python-decouple==3.8
pytz==2022.7.1
requests==2.30.0
rest-framework-simplejwt==0.0.2
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
yarl==1.9.2
